refactor: 로고 png에서 svg로 교체 및 여러 곳에서 쓰이던 이미지 하나로 통일
