+++ src/components/common/Header.scss
@@ -26,6 +26,8 @@
 }
 
 .logo {
+  width: 170px;
+  height: 32px;
   grid-column: 2;
   justify-self: center;
   cursor: pointer;
